count-objects: report statistics about kept packs
Specifically: number of kept packs, size of kept packs (and indexes),
and number of objects in kept packs.

Add documentation of resulting fields.  While here, correct the omission
from ae72f68541 ("count-objects -v: show number of packs as well.",
2006-12-27) of the "packs" field.

Update tests which were searching for the "in-pack" field without
the terminating colon and so were now also picking up the "in-pack-keep"
field.

Signed-off-by: Nathaniel Filardo <nwf20@cl.cam.ac.uk>